feat(config): split SettingsEditor into module, add Auto Mode, Plugins, Sandbox filesystem
Phase 0 — Split the 1,090-line SettingsEditor.tsx into a module:
- settings/SettingsEditor.tsx — orchestrator (~200 lines)
- settings/constants.ts — option arrays
- settings/field-components.tsx — SwitchSetting, ListEditor, etc.
- settings/types.ts — SettingsCardProps interface
- settings/cards/*.tsx — 11 existing cards extracted into individual files

Phase 1 — Auto Mode card (new):
- autoMode.environment, .allow, .soft_deny (ListEditors)
- permissions.disableAutoMode toggle
- 'auto' added to PERMISSION_MODE_OPTIONS
- Scope guard: hidden in project scope (user/local/managed only)

Phase 2 — Plugin Management card (new):
- enabledPlugins (ListEditor)
- extraKnownMarketplaces (ObjectArrayEditor — new field component)
- Managed-only: blockedMarketplaces, pluginTrustMessage, strictKnownMarketplaces

Phase 3 — Sandbox filesystem rules:
- sandbox.filesystem.allowRead/denyRead (ListEditors, gitignore patterns)
- Managed-only: allowManagedReadPathsOnly, allowManagedDomainsOnly
