+++ frontend/src/features/config/settings/cards/AttributionCard.tsx
@@ -1,0 +1,33 @@
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
+import { AttributionField } from '../field-components'
+import type { SettingsCardProps } from '../types'
+
+export function AttributionCard({ getSettingRaw, updateSetting }: SettingsCardProps) {
+  return (
+    <Card>
+      <CardHeader>
+        <CardTitle>Attribution</CardTitle>
+        <CardDescription>Configure commit and PR attribution text</CardDescription>
+      </CardHeader>
+      <CardContent className="space-y-4">
+        <AttributionField
+          id="attributionCommit"
+          label="Commit Attribution"
+          description="Text appended to commit messages"
+          defaultText="Co-Authored-By: Claude <noreply@anthropic.com>"
+          rawValue={getSettingRaw('attribution.commit')}
+          onChange={(v) => updateSetting('attribution.commit', v)}
+        />
+
+        <AttributionField
+          id="attributionPr"
+          label="PR Attribution"
+          description="Text appended to pull request descriptions"
+          defaultText="Generated by Claude"
+          rawValue={getSettingRaw('attribution.pr')}
+          onChange={(v) => updateSetting('attribution.pr', v)}
+        />
+      </CardContent>
+    </Card>
+  )
+}
