+++ frontend/src/features/config/settings/index.ts
@@ -1,0 +1,1 @@
+export { SettingsEditor } from './SettingsEditor'
